Fix critest "runtime should support attach" test
The conformance CRI test does not stop the running container after
checking its output. It expects that the connection gets dropped when
there is no more input to read. We now adjust the current attach
behavior to stop in that case, which hopefully does not break the Podman
use case.

See: https://github.com/kubernetes-sigs/cri-tools/blob/909f652/pkg/validate/streaming.go#L266-L305

We also strip down the size of the data packets to return the right
amount of received bytes in the client.

A debug message has been changed which correctly logs the amount of
written packets.

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>