Reduce memory usage for attach
It's unrealistic that we have to buffer 1000 packets, especially since
they should be consumed by in a fast manner. We now still keep a safe
zone of 4 packets in the channel, but this already reduces the memory by
multiple hundred kilobytes.

We also optimize the release build by stripping of the binary as well as
removing the debug symbols per default.

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>