Remove attach socket path if not in use
We cannot really know if the unix listener is still in use from the
conmon-rs perspective, because it could be possible that conmonrs got
killed while still having readers open on the client side.

We now add a cleanup logic to the client to remove the socket path if no
readers are available any more. This allows conmonrs to still create
multiple socket paths in parallel and keep them open until all clients
are gone, the container terminates or the server does a shutdown.

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>