Fix OOM watcher for cgroupv2 `oom_kill` events
It may be possible that the container process already got signaled but
is still running, whereas we will not get a `modify` (but `other`) event
on the file watcher. Beside that, it will also not update the `oom`
entry of `memory.events`, but the `oom_kill`, which provides another
indicator for a possible out of memory kill.

Ref: https://www.kernel.org/doc/Documentation/cgroup-v2.txt

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>