Use default umask `0022`
Under non-reproducible circumstances it can happen that the umask of
conmon-rs is set to 0000, which can be a security issue. To avoid that,
we now use a default umask and set it directly in the init module.

We had a similar fix in CRI-O in cri-o/cri-o#5904, but we have to do it
here again because conmon-rs is a direct parent of PID 1.

Refers to https://issues.redhat.com/browse/OCPBUGS-8057

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>